requirements.txt
sqlalchemy
MySQL-python
PyYAML
ipdb